Refactor ApiReference and Schema components
