+++ gatsby/src/pages/api-reference/other.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import data from '../../../static/specs/minfraud.json';
-import ApiSpec from '../../templates/ApiSpec';
+import ApiSpec from '../../templates/ApiReference';
 
 const Minfraud = (): React.ReactElement => (
   <ApiSpec
